+++ src/gui.js
@@ -87,7 +87,7 @@ HatBlockMorph, ZOOM*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2026-March-02';
+modules.gui = '2026-March-10';
 
 // Declarations
 
@@ -5168,6 +5168,15 @@ IDE_Morph.prototype.settingsMenu = function () {
         'uncheck to show all primitive block categories',
         'check to hide empty primitive block categories',
         false
+    );
+    addPreference(
+        'Enforce input types',
+        () => ScriptsMorph.prototype.enforceTypes =
+            !ScriptsMorph.prototype.enforceTypes,
+        ScriptsMorph.prototype.enforceTypes,
+        'uncheck to allow\ndropping reporters\ninto unmatching slots',
+        'disable dropping reporters\ninto unmatching slots',
+        true
     );
     addPreference(
         'Wrap list indices',